TiledMap Update - BackgroundMap has been deleted in favor of TiledMap. TiledMap   uses the bridge pattern to separate implementation from the   interface. This change allows for subclassing TiledMap to extend   functionality. Previously, extension was done through composition   and delegation, but this could have been a maintenance hazard   were the original class ever to be modified. - This change renders many classes in this snapshot obsolete,   including DoodadMap and BackgroundMapFactory. - TiledMap does not strictly depend on Background, allowing for   custom backgrounds in the future. - TiledMap and CoordConverter consider size as a long type. This   allows height and width to be their appropriate maximum   capacities without penalty.

+++ src/util/CoordConverter.cpp
@@ -1,9 +1,9 @@
 #include "CoordConverter.h"
 
-unsigned int CoordConverter::coordsToIndex(const unsigned int& x,
+unsigned long CoordConverter::coordsToIndex(const unsigned int& x,
                                            const unsigned int& y,
 					   const unsigned int& width) {
-  unsigned int ret = ((y * width) + x);
+  unsigned long ret = ((y * width) + x);
   return ret;
 }
